feat: 회원 리뷰 CRUD 기능 구현
- 리뷰 0.5 단위를 체크하기 위해 어노테이션 추가
- 리뷰 사진 최대 6개까지 넣을 수 있도록 검증
- 사용자가 남긴 리뷰를 확인할 수 있도록 구현
- NPE 방지 코드 추가
- removeReviewScores 방어 로직 추가
- review.getPlace().getId() 호출 시 Lazy Loading 예외 발생 위험 해결
- @Modifying에 clearAutomatically = true 추가

+++ src/test/java/io/dnd/goyo/domain/place/service/PlaceRecommendationIntegrationTest.java
@@ -527,7 +527,7 @@ class PlaceRecommendationIntegrationTest {
                 .build());
     }
 
-    private Review createReview(User user, Place place, int rating) {
+    private Review createReview(User user, Place place, double rating) {
         Review review = Review.create(user, place, rating, null, null, null, null, null, null);
         ReflectionTestUtils.setField(review, "status", ReviewStatus.ACTIVE);
         return reviewRepository.save(review);
